requirements.txt
llama-index
llama-index-llms-openai
llama-index-llms-gemini
llama-index-embeddings-ollama
openai
python-dotenv
Flask
